Implement compute shader morphs

+++ src/renderer/model/mod.rs
@@ -25,8 +25,8 @@ pub trait Model {
 	fn try_clone(&self, renderer: &mut Renderer) -> Result<Box<dyn Model>, ModelError>;
 }
 
-pub trait VertexIndex: Index + Copy + Send + Sync + Sized + Into<usize> + 'static {}
-impl<T> VertexIndex for T where T: Index + Copy + Send + Sync + Sized + Into<usize> + 'static {}
+pub trait VertexIndex: Index + Copy + Send + Sync + Sized + Into<i32> + 'static {}
+impl<T> VertexIndex for T where T: Index + Copy + Send + Sync + Sized + Into<i32> + 'static {}
 
 #[derive(Debug, Error)]
 pub enum ModelError {
@@ -44,5 +44,7 @@ pub enum ModelError {
 pub enum ModelRenderError {
 	#[error(display = "{}", _0)] DrawIndexedError(#[error(source)] command_buffer::DrawIndexedError),
 	#[error(display = "{}", _0)] CopyBufferError(#[error(source)] command_buffer::CopyBufferError),
+	#[error(display = "{}", _0)] FillBufferError(#[error(source)] command_buffer::FillBufferError),
+	#[error(display = "{}", _0)] DispatchError(#[error(source)] command_buffer::DispatchError),
 	#[error(display = "{}", _0)] DeviceMemoryAllocError(#[error(source)] memory::DeviceMemoryAllocError),
 }
